+++ httomo/method_wrappers/rotation.py
@@ -1,5 +1,5 @@
+from httomo.block_interfaces import T, Block
 from httomo.method_wrappers.generic import GenericMethodWrapper
-from httomo.runner.dataset import DataSetBlock
 from httomo.runner.method_wrapper import MethodParameterDictType
 from httomo.runner.methods_repository_interface import MethodRepository
 from httomo.utils import Pattern, catchtime, log_once, xp
@@ -62,7 +62,7 @@ class RotationWrapper(GenericMethodWrapper):
     def _build_kwargs(
         self,
         dict_params: MethodParameterDictType,
-        dataset: Optional[DataSetBlock] = None,
+        dataset: Optional[Block] = None,
     ) -> Dict[str, Any]:
         assert (
             dataset is not None
@@ -100,7 +100,7 @@ class RotationWrapper(GenericMethodWrapper):
             )
             return None
 
-    def _run_method(self, block: DataSetBlock, args: Dict[str, Any]) -> DataSetBlock:
+    def _run_method(self, block: T, args: Dict[str, Any]) -> T:
         """Different CoR estimation methods require different inputs, e.g., the VoCentering
         method needs the full sinogram, while the PhaseCorrelation method needs specific projections.
         """
@@ -207,7 +207,7 @@ class RotationWrapper(GenericMethodWrapper):
             sino -= darks1d / denom
         return sino[:, np.newaxis, :]
 
-    def _process_return_type(self, ret: Any, input_block: DataSetBlock) -> DataSetBlock:
+    def _process_return_type(self, ret: Any, input_block: T) -> T:
         if type(ret) == tuple:
             # cor, overlap, side, overlap_position - from find_center_360
             self._side_output["cor"] = float(ret[0])
